chore: centralize round KPI data
