+++ apps/balancer-tools/src/app/apr/(components)/OverviewCards.tsx
@@ -1,9 +1,24 @@
+import { formatDate } from "@bleu-balancer-tools/utils";
+
+import { Round } from "../(utils)/rounds";
 import { KPI } from "./KPI";
 
 interface CardDetail {
   title: string;
   content: JSX.Element | string;
   tooltip?: string;
+}
+
+export function getRoundDetails(roundId: string) {
+  const round = Round.getRoundByNumber(roundId);
+  return [
+    { title: "Round Number", content: roundId },
+    {
+      title: "Round Ended",
+      content: formatDate(round.endDate),
+      tooltip: "Every round ends on a Thursday at 00:00 UTC",
+    },
+  ];
 }
 
 function OverviewCards({ cardsDetails }: { cardsDetails: CardDetail[] }) {
