+++ apps/balancer-tools/src/app/apr/round/(components)/RoundOverviewCards.tsx
@@ -1,9 +1,9 @@
-import { formatDate } from "@bleu-balancer-tools/utils";
-
 import * as balEmissions from "#/lib/balancer/emissions";
 import { formatNumber } from "#/utils/formatNumber";
 
-import OverviewCards from "../../(components)/OverviewCards";
+import OverviewCards, {
+  getRoundDetails,
+} from "../../(components)/OverviewCards";
 import { getBALPriceByRound } from "../../(utils)/getBALPriceByRound";
 import { Round } from "../../(utils)/rounds";
 
@@ -25,12 +25,7 @@ export default async function RoundOverviewCards({
         balEmissions.weekly(round.endDate.getTime() / 1000),
       ),
     },
-    { title: "Round Number", content: roundId },
-    {
-      title: "Round Ended",
-      content: formatDate(round.endDate),
-      tooltip: "Every round ends on a Thursday at 00:00 UTC",
-    },
+    ...getRoundDetails(roundId),
   ];
   return <OverviewCards cardsDetails={cardsDetails} />;
 }
